Implement TTL mechanism and integrate into server loop
- Add handling for CMD_EXPIRE in the server's request processing.
- Update epoll_loop to use dynamic timeouts based on the next TTL
expiration and call ttl_process_expirations.
- Refine TTL list management in src/ttl.c to safely handle re-expiring
keys and prevent list corruption.
- Update node structure to use long long for absolute expiration
timestamps.
- Ensure node memory is zero-initialized using calloc.

+++ include/database.h
@@ -10,7 +10,7 @@ typedef struct LRU_list_Node LRU_list_Node;
 typedef struct node{
   ht_entry entry;
   bool is_file;
-  uint32_t ttl;
+  long long ttl;
     //LRU node 
   LRU_list_Node* LRU_node;
   TAILQ_ENTRY(node) ttl_link;
